# openapi.yaml
openapi: 3.0.3
info:
  title: Tennis Court Finder API
  description: |
    Monitors tennis court availability across integrated booking systems and sends
    real-time alerts to users when requested court times become available.

    ## Authentication Flow
    1. User requests a one-time password (OTP) sent to their email
    2. User verifies the OTP
    3. Server returns a JWT stored in an HTTP-only cookie
    4. All authenticated endpoints read identity from the JWT cookie
  version: 0.1.0
  contact:
    name: Tennis Court Finder
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request a one-time password sent to the given email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          description: Too many OTP requests – try again later
  /api/auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and receive a JWT session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: OTP verified – JWT set in cookie
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session=eyJhbGciOi...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Clear the session cookie
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user info
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Clubs ───────────────────────────────────────────────────────────────
  /api/clubs:
    get:
      operationId: listClubs
      tags: [clubs]
      summary: List all integrated tennis clubs
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
          description: Filter clubs by city name (case-insensitive partial match)
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of clubs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubListResponse"
  /api/clubs/{club_id}:
    get:
      operationId: getClub
      tags: [clubs]
      summary: Get details of a specific club
      parameters:
        - $ref: "#/components/parameters/ClubIdParam"
      responses:
        "200":
          description: Club details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Courts ──────────────────────────────────────────────────────────────
  /api/clubs/{club_id}/courts:
    get:
      operationId: listCourts
      tags: [courts]
      summary: List courts for a club
      parameters:
        - $ref: "#/components/parameters/ClubIdParam"
        - name: surface_type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/SurfaceType"
          description: Filter by surface type
        - name: court_type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/CourtType"
          description: Filter by indoor/outdoor
      responses:
        "200":
          description: List of courts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Court"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/clubs/{club_id}/courts/{court_id}:
    get:
      operationId: getCourt
      tags: [courts]
      summary: Get details of a specific court
      parameters:
        - $ref: "#/components/parameters/ClubIdParam"
        - $ref: "#/components/parameters/CourtIdParam"
      responses:
        "200":
          description: Court details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Court"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Time Slots ──────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: JWT session token stored in an HTTP-only cookie

  # ── Reusable Parameters ──────────────────────────────────────────────
  parameters:
    ClubIdParam:
      name: club_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        example: seb-arena
      description: URL-friendly slug identifier of the tennis club
    CourtIdParam:
      name: court_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the court
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    PageSizeParam:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page

  # ── Reusable Responses ──────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ── Schemas ─────────────────────────────────────────────────────────
  schemas:
...
    SurfaceType:
      type: string
      enum: [hard, clay, carpet, grass, artificial_grass]
      description: Court surface material
    CourtType:
      type: string
      enum: [indoor, outdoor]
      description: Whether the court is indoors or outdoors
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: Resource not found
    ValidationErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
    PaginationMeta:
      type: object
      required: [page, page_size, total_items, total_pages]
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 42
        total_pages:
          type: integer
          example: 3

    # ── Auth ────────────────────────────────────────────────────────
    OtpRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: player@example.com
    OtpRequestResponse:
      type: object
      required: [message, expires_in_seconds]
      properties:
        message:
          type: string
          example: OTP sent to your email
        expires_in_seconds:
          type: integer
          example: 300
    OtpVerifyRequest:
      type: object
      required: [email, otp_code]
      properties:
        email:
          type: string
          format: email
          example: player@example.com
        otp_code:
          type: string
          minLength: 6
          maxLength: 6
          example: "482910"
    AuthResponse:
      type: object
      required: [message, user]
      properties:
        message:
          type: string
          example: Authenticated successfully
        user:
          $ref: "#/components/schemas/UserInfo"
    UserInfo:
      type: object
      required: [email, created_at]
      properties:
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time

    # ── Club ────────────────────────────────────────────────────────
    Club:
      type: object
      required: [id, name, address, city]
      properties:
        id:
          type: string
          description: URL-friendly slug identifier
          example: seb-arena
        name:
          type: string
          example: City Tennis Club
        address:
          type: string
          example: 123 Racket Lane
        city:
          type: string
          example: Vilnius
        phone:
          type: string
          nullable: true
          example: "+37060000000"
        website:
          type: string
          format: uri
          nullable: true
          example: https://citytennisclub.lt
        image_url:
          type: string
          format: uri
          nullable: true
        courts_count:
          type: integer
          description: Number of courts at this club
          example: 6
    ClubListResponse:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Club"
        meta:
          $ref: "#/components/schemas/PaginationMeta"

    # ── Court ───────────────────────────────────────────────────────
    Court:
      type: object
      required: [id, club_id, name, surface_type, court_type]
      properties:
        id:
          type: string
          format: uuid
        club_id:
          type: string
          description: Slug of the parent club
          example: seb-arena
        name:
          type: string
          example: Court 1
        surface_type:
          $ref: "#/components/schemas/SurfaceType"
        court_type:
          $ref: "#/components/schemas/CourtType"
        description:
          type: string
          nullable: true
          example: Main show court with floodlights

    # ── Time Slot ───────────────────────────────────────────────────
...